将无关类删除，整合到Lib类中 需要通过命令行执行了
1. 整合所有方法到Lib类中 之前那么多类其实都是无意义的 陋习罢了
2. 现在需要通过命令行先编译，然后传参数才能执行了
