--- .gitignore
@@ -1,14 +1,0 @@
-/221801310/.idea/
-/.git/
-/.gitignore/
-*.iml
-/.idea/
-/221801310/src/input.txt
-/out/
-/221801310/src/output.txt
-/.gradle/
-/221801310/production/
-/221801310/Test/
-/221801310/src/测试文件.zip
-/221801310/src/input7.txt
-/221801310/src/测试文件/
