--- 221801310/src/CountAsciiChar.java
@@ -1,11 +1,0 @@
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
-
-public class CountAsciiChar{
-
-    /*统计文件ascii字符数量 传入参数text为文本内容字符串*/
-    public static int countChar(String text){
-        return text.length();
-    }
-
-}
