+++ 221801310/src/WordCount.java
@@ -1,3 +1,4 @@
+import java.io.File;
 import java.util.HashMap;
 import java.util.LinkedHashMap;
 import java.util.List;
@@ -7,26 +8,30 @@ import java.util.concurrent.*;
 /*主函数,通过命令行调用，读取和输出的文件名将在命令行参数中给出*/
 public class WordCount{
     public static void main(String[] args){
+        File inputFile;
+        File outputFile;
+        //通过命令行给出文件路径
+       if(args.length!=2){
+           System.out.println("请输入两个文件的绝对路径或相对路径！");
+           return;
+       }
+       inputFile = new File(args[0]);
+       outputFile = new File(args[1]);
+
         final CountDownLatch countDownLatch = new CountDownLatch(2);
         FutureTask<HashMap<String, Integer>> futureTask1;
         FutureTask<Integer> futureTask2;
-        //开始计时
-        long startTime = System.currentTimeMillis();
 
-        String inputFile = "input7.txt";
-        String outputFileName = "out7.txt";
         String content;     //文本内容
         long asciiCharNum;
         long wordNum;
         long lines = 0;
         HashMap<String, Integer> legalWords = new LinkedHashMap<>();
 
-        content = ReadTxt.readTxt(inputFile);
-        asciiCharNum = CountAsciiChar.countChar(content);
-
-        //合法单词的hashMap
-//        legalWords = SplitWord.findLegal(content);
-        Callable callable1 = new Callable() {
+        content = Lib.readTxt(inputFile);
+        asciiCharNum = Lib.countChar(content);
+        //线程1
+        Callable<HashMap<String, Integer>> callable1 = new Callable<HashMap<String, Integer>>() {
             HashMap<String, Integer> legalWords;
             @Override
             public HashMap<String, Integer> call() {
@@ -35,22 +40,20 @@ public class WordCount{
                 }catch (Exception e){
                     e.printStackTrace();
                 }
-                this.legalWords = SplitWord.findLegal(content);
+                this.legalWords = Lib.findLegal(content);
                 countDownLatch.countDown();
-                return legalWords;
+
+                return this.legalWords;
             }
         };
         futureTask1 = new FutureTask<HashMap<String, Integer>>(callable1);
         new Thread(futureTask1).start();     //创建新线程
-
-
-//        lines = CountLine.countLine(content);
-        Callable callable2 = new Callable() {
+        //线程2
+        Callable<Integer> callable2 = new Callable<Integer>() {
             Integer line;
             @Override
             public Integer call() {
-                line = CountLine.countLine(content);
-//                line = CountLine.countLine1(inputFile);
+                line = Lib.countLine(content);
                 countDownLatch.countDown();
                 return line;
             }
@@ -58,7 +61,6 @@ public class WordCount{
         futureTask2 = new FutureTask<Integer>(callable2);
         new Thread(futureTask2).start();
 
-        //取出结果
         try {
             lines = futureTask2.get();
             legalWords = futureTask1.get();
@@ -70,25 +72,14 @@ public class WordCount{
             e.printStackTrace();
         }
 
-//        hashMap = CountFrequency.countFrequency(content);
-//        hashMap = CountFrequency.countFrequency(legalWords);
         try {
-//            System.out.println("主线程"+Thread.currentThread().getName()+"等待子线程执行完成...");
             countDownLatch.await(); //阻塞当前主线程
-//            System.out.println("主线程"+Thread.currentThread().getName()+"开始执行...");
         }
         catch (InterruptedException e){
             e.printStackTrace();
         }
-        wordNum = SplitWord.countWordNum();
-        List<Map.Entry<String, Integer>> list = CountFrequency.sortHashMap(legalWords);
-
-
-        OutputToTxt.outputToTxt(asciiCharNum, wordNum, lines, list, outputFileName);
-
-        long endTime = System.currentTimeMillis();
-        System.out.println("程序运行总时间："+(endTime-startTime)+"ms");
+        wordNum = Lib.countWordNum();
+        List<Map.Entry<String, Integer>> list = Lib.sortHashMap(legalWords);
+        Lib.outputToTxt(asciiCharNum, wordNum, lines, list, outputFile);
     }
 }
-
-
